problem with find and id
