+++ views/index.erb
@@ -17,8 +17,8 @@
        </td>
        <td class="newuser" style="width 15%">
          <% if @user %>
-  <h1>Welcome, <%= @user.email %></h1>
-<% end %>
+            <h1>Welcome, <%= @user.email %></h1>
+         <% end %>
        </td>
      </tr>
    </table>
